+++ src/language/dafnyInstallation.ts
@@ -4,40 +4,29 @@ import { ConfigurationConstants } from '../constants';
 import Configuration from '../configuration';
 import { Executable } from 'vscode-languageclient/node';
 import { GitHubReleaseInstaller } from './githubReleaseInstaller';
-import { getCliPath } from './cliCopier';
-import { getDotnetExecutablePath } from '../dotnet';
+import { CustomPathInstaller } from './customPathInstaller';
 import * as os from 'os';
 import { FromSourceInstaller } from './fromSourceInstaller';
 
 export class DafnyInstaller {
+  private readonly customPathInstaller: CustomPathInstaller;
+
   public constructor(
     public readonly context: ExtensionContext,
     public readonly statusOutput: OutputChannel
-  ) {}
+  ) {
+    this.customPathInstaller = new CustomPathInstaller(this.context, this.statusOutput);
+  }
 
   public async getCliExecutable(server: boolean, newArgs: string[], oldArgs: string[]): Promise<Executable> {
-    const { path: dotnetExecutable } = await getDotnetExecutablePath();
-    const configuredCliPath = await getCliPath(this.context);
-    if(configuredCliPath) {
-      if(server) {
-        newArgs.unshift('server');
-      }
-      if(configuredCliPath.endsWith('.dll')) {
-        return {
-          command: dotnetExecutable,
-          args: [ configuredCliPath ].concat(newArgs)
-        };
-      } else {
-        return {
-          command: configuredCliPath,
-          args: newArgs
-        };
-      }
+    const customExecutable = await this.customPathInstaller.getExecutable(this.context, server, newArgs, oldArgs);
+    if(customExecutable) {
+      return customExecutable;
     }
 
     if(os.type() === 'Darwin' && os.arch() !== 'x64') {
       // Need to build from source and move all files from Binary/ to the out/resource folder
-      const sourceInstaller = new FromSourceInstaller(new GitHubReleaseInstaller(this.context, this.statusOutput));
+      const sourceInstaller = new FromSourceInstaller(this.context, this.statusOutput);
       return await sourceInstaller.getExecutable(server, newArgs, oldArgs);
     }
     return await new GitHubReleaseInstaller(this.context, this.statusOutput).getExecutable(server, newArgs, oldArgs);
